+++ framework/db/oci/Command.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace yii\db\oci;
+
+class Command extends \yii\db\Command
+{
+    /**
+     * {@inheritdoc}
+     */
+    protected function bindPendingParams()
+    {
+        foreach ($this->_pendingParams as $name => $value) {
+            if (\PDO::PARAM_STR === $value[1]) {
+                $this->pdoStatement->bindParam($name, $value[0], $value[1], strlen($value[0]));
+            } else {
+                $this->pdoStatement->bindValue($name, $value[0], $value[1]);
+            }
+        }
+        $this->_pendingParams = [];
+    }
+
+}