+++ framework/CHANGELOG.md
@@ -4,6 +4,7 @@ Yii Framework 2 Change Log
 2.0.33 under development
 ------------------------
 
+- Bug #17679: Fix oracle exception "ORA-01461: can bind a LONG value only for insert into a LONG column" when inserting 4k+ string with oracle (vinpel)
 - Bug #17850: Update to `ReplaceArrayValue` config exception message (alex-code)
 - Bug #17843: Fix `yii\web\Session::setCookieParamsInternal` check param "samesite" (schevgeny)
 - Bug #17828: Fix `yii\web\UploadedFile::saveAs()` failing when error value in `$_FILES` entry is a string (haveyaseen)